refactor: define supported frontend version
