+++ src/lpsim/network/__version__.py
@@ -1,0 +1,12 @@
+# This file will record minimum supported version of frontend and current version.
+# When frontend is updated, modify the version here.
+from .. import __version_tuple__, __version__
+
+__frontend_version__ = "0.4.5.0"
+
+
+__all__ = [
+    "__frontend_version__",
+    "__version__",
+    "__version_tuple__",
+]
