+++ src/lpsim/network/http_room_server.py
@@ -14,7 +14,7 @@ from starlette.middleware.gzip import GZipMiddleware
 from multiprocessing import Process, Queue
 
 from .http_server import HTTPServer
-from .. import __version_tuple__, __version__  # type: ignore
+from .__version__ import __version_tuple__, __version__, __frontend_version__
 
 
 class OneRoomWorker(Process):
@@ -169,12 +169,19 @@ class HTTPRoomServer:
         async def get_version():
             """
             Return the version of lpsim.
+
+            Return:
+                version: version string
+                version_tuple: version tuple
+                support_version: support version of frontend
+                info: information of the server
             """
             return {
                 "version": __version__,
                 "version_tuple": __version_tuple__,
+                "support_version": __frontend_version__,
                 "info": {
-                    "class": "HTTPRoomServer",
+                    "class": "HTTPServer",
                 },
             }
